+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## x.y.z (yyyy-mm-dd)
+
+  - Support adding sub-queries to QueryBuilder.
+
 ## 0.5.2 (2013-09-26)
 
   - Support native XML-types of databases.